pip-requirements.txt
django-jenkins==0.110.0
pep8==1.7.1
pyflakes==2.2.0
python-dotenv==0.15.0
comtypes==1.1.7
docx==0.2.4
docx-mailmerge==0.4.0
pywin32==300
requests==2.25.1
mkdocs = 1.1.2
mkdocs-gitbook